requirements.txt
loguru
pycryptodome
requests
pydash